Fixed issue where strategy always did first move
Timeline of nextMove() made it so that the record of opponent never got updated. Moved placement of update to happen before checking if first move. Also changed check from the 0th entry to the 1st, since now the default opponent last move wil be placed at the 0th position on start of a prisoner's dilmma run.