Changed logic for 0 record and first run
Previously, if the record length was 0 or if we had nothing stored in our record (first run), the strategy would always do the first move. Changed the logic so that instead the first move is stored and the wildcard chance is calculated. If true, the move is flipped and then returned.